+++ frontend/src/employees/Employee.js
@@ -1,0 +1,40 @@
+import React from 'react'
+import EmployeeForm from './EmployeeForm'
+import useUserData from '../hooks/useUserData'
+
+export default function Worksheet() {
+  const {
+    HandleData,
+    handleClick,
+    popUpMessage,
+    sent,
+    userData,
+  } = useUserData()
+
+  return window.location.pathname === '/employees/new' ? (
+    <>
+      <EmployeeForm
+        handleClick={handleClick}
+        sent={sent}
+        popUpMessage={popUpMessage}
+        sendData={HandleData}
+        title="Új munkatárs létrehozása"
+        registration={false}
+      />
+    </>
+  ) : (
+    <>
+      {userData && (
+        <EmployeeForm
+          handleClick={handleClick}
+          sent={sent}
+          popUpMessage={popUpMessage}
+          sendData={HandleData}
+          title="Adatok módosítása"
+          user={userData}
+          registration={true}
+        />
+      )}
+    </>
+  )
+}
